% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup a standardized folder and file structure for a research analysis
project.}
\usage{
setup_project(
  path,
  type = c("github", "gitlab"),
  add_gitignore = TRUE,
  gitignore_templates = c("r", "macos")
)
}
\arguments{
\item{path}{A path to a new directory.}

\item{type}{chr. Vector specifying whether this project will be hosted on
github or gitlab. Default is github.}

\item{add_gitignore}{logical. Specify whether gitignore should be added.
Default is TRUE.}

\item{gitignore_templates}{chr. A character vector using values included in
\code{\link[gitignore]{gi_available_templates}}. Default is "r" and
"macos".}
}
\value{
Project setup with folders and files necessary for a standard
research project.
}
\description{
This starts the project by setting up a common folder and file
infrastructure,as well as adding some useful files to start the project.
}
\examples{
\dontrun{
# Use a temporary location
new_proj_name <- fs::path_temp("Analysis")
setup_project(new_proj_name)
}
}
