% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_index.R
\name{update_index}
\alias{update_index}
\title{Update index table}
\usage{
update_index(
  html_folder = "docs",
  include_number = TRUE,
  include_description = TRUE
)
}
\arguments{
\item{html_folder}{Character. The folder containing \code{.html}s to be referenced
in the contents table. This path should be relative to the root folder of
the R project (i.e. the folder containing the \code{.Rproj} file). Default is
"docs/".}

\item{include_number}{boolean. Whether to include the numbering of the
analysis i.e. if the document has been labelled with a number to indicate
ordering (e.g. "\verb{01_example.rmd}"). If no numbering requested, analyses
will be ordered alphabetically. Default is TRUE.}

\item{include_description}{boolean. Whether to include a description of the
analysis, which will be taken straight from the \code{.rmd}. This function
assumes that that this description is described in the section preceded by
"> Aim:" and ended with "<br><br>", as in the example analysis templates
available in both templates. Both of these templates were generated using:
https://github.com/RHReynolds/rmdplate. Default is TRUE.}
}
\value{
a \code{\link[tibble:tbl_df-class]{tibble}}, but in md formatting.
}
\description{
\code{update_index} will generate a \code{\link[tibble:tbl_df-class]{tibble}}
object and display it with md formatting. It will contain the title of the
analysis, as extracted from the \code{.rmd} with a link to the corresponding
\code{.html}. If chosen, it will also show the number of the analysis and a
description of the analysis, as extracted from the file name and \code{.rmd},
respectively.
}
